+++ framework/Database/ResultInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Framework\Database;
+
+interface ResultInterface
+{
+    public function fetch(): array|false;
+}